% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cleanCorpus.R
\name{AddWordCombinations}
\alias{AddWordCombinations}
\title{Prepares combinations of words to be changed}
\usage{
AddWordCombinations(originalCombination = "", toBeUsed = "",
  nameOfProject = "", dfedit = FALSE)
}
\arguments{
\item{newStopwords}{A character vector of words.}
}
\value{
originalCombination, toBeUsed A character vector.of length 1.
}
\description{
Prepares combinations of words to be changed.
}
\examples{
originalCombination <- "European Union"
toBeUsed <- "europeanunion"
wordCombinations <- AddWordCombinations(originalCombination, toBeUsed, nameOfProject)
}

