% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyAnalyseDataset.R
\name{AnalyseDataset}
\alias{AnalyseDataset}
\title{Starts a Shiny app that facilitates analysing a `castarter` dataset through time series.}
\usage{
AnalyseDataset()
}
\description{
Starts a Shiny app that facilitates analysing a `castarter` dataset through time series.
}
\examples{
\dontrun{
AnalyseDataset()
}
}
