% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyseTerms.R
\name{AnalyseTerms}
\alias{AnalyseTerms}
\title{Compares frequency of specific words across websites.}
\usage{
AnalyseTerms(corpus, project, specificTerms, mode = "graph",
  includeOnly = NULL, order = TRUE, tipology = NULL,
  frequency = "relative", corpusDtm = NULL)
}
\arguments{
\item{corpus}{A TM corpus.}

\item{project}{Name of 'castarter' project. Must correspond to the name of a folder in the current working directory.}

\item{specificTerms}{Character vector of terms to be compared.}

\item{corpusDtm}{A document-term matrix. If provided, computation is faster.}
}
\description{
To be incorporated into ShowFreq
}
\examples{
AnalyseTerms(corpus, project, specificTerms)
}
