% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cleanCorpus.R
\name{CleanCorpus}
\alias{CleanCorpus}
\title{Calls various 'tm' functions to clean up the corpus.}
\usage{
CleanCorpus(corpus, stripWhitespace = TRUE, toLowerCase = TRUE,
  removeNumbers = TRUE, removePunctuation = TRUE,
  removeControlCharacters = TRUE, removeCharacter = "")
}
\arguments{
\item{corpus}{A corpus as created by the 'tm' package including metadata.}
}
\value{
A corpus as created by the 'tm' package including metadata, after all enabled transformations are applied.
}
\description{
It applies to a corpus common operations to prepare it for further analysis using functions of the 'tm' package.
}
\examples{
corpus <- CleanCorpus(corpus)
}
\keyword{tm}

