% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanCorpus.R
\name{CombineWords}
\alias{CombineWords}
\title{Combine words in a corpus}
\usage{
CombineWords(corpus, wordCombinations)
}
\arguments{
\item{corpus}{A corpus created with the 'tm' package.}
}
\value{
A corpus with words combined.
}
\description{
Combine words in a corpus according to word combinations provided by the user..
}
\examples{
corpus <- CombineWords(corpus, wordCombinations)
}
