% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareCorpus.R
\name{ConvertToCorpus}
\alias{ConvertToCorpus}
\title{Converts a 'castarter' dataset into a 'tm' corpus.}
\usage{
ConvertToCorpus(dataset, quanteda = FALSE)
}
\arguments{
\item{dataset}{A data.frame created by 'castarter' including metadata and full text articles to be converted into a corpus.}

\item{Logical, }{defaults to FAlSE. If TRUE generates corpus of the 'quanteda' package, otherwise of the 'tm' package.}
}
\value{
A corpus as created by the 'tm' or 'quanteda' package including metadata.
}
\description{
Takes a dataset created with 'castarter' and converts it into a 'tm' corpus. Metadata are automatically imported.
}
\examples{
corpus <- ConvertToCorpus(dataset)
}

