% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanup.R
\name{CountHtml}
\alias{CountHtml}
\title{Counts the number of HTML found in all folders of `castarter` projects}
\usage{
CountHtml(exportCsv = FALSE)
}
\arguments{
\item{exportCsv}{Defaults to FALSE. If TRUE, exports the outputs as csv file in the base folder (usually `castarter`, inside the working folder).}
}
\value{
A data frame with the number of html files inlcuded in project subfolders
}
\description{
This function facilitates the identification of folders with high numbers of Html files. Particularly useful before backups.
}
\examples{
CountHtml()
}
