% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countWords.R
\name{CountWords}
\alias{CountWords}
\title{Count total words by date and website}
\usage{
CountWords(corpusDtm, byWebsite = TRUE, byYear = TRUE, byDay = FALSE)
}
\arguments{
\item{corpusDtm}{A document-term matrix created with the 'TM' package or a document-feature matrix of the 'quanteda' type.}

\item{byDay}{Logical, defaults to FALSE. If TRUE, overrides byYear.}
}
\value{
A data.frame
}
\description{
Count total words by date and website
}
\examples{
totalWords <- CountWords(corpusDtm)
}

