% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_CreateDataset.R
\name{CreateDataset}
\alias{CreateDataset}
\title{Starts app that facilitates creation of a 'castarter' dataset}
\usage{
CreateDataset()
}
\description{
Starts app that facilitates creation of a 'castarter' dataset
}
\examples{
CreateDataset()
}
