% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractMetadata.R
\name{CreateDatasetFromHtml}
\alias{CreateDatasetFromHtml}
\title{Extracts metadata and text from local html files}
\usage{
CreateDatasetFromHtml(nameOfProject, nameOfWebsite, articlesLinks = NULL,
  dateFormat = "dBY", ExtractDatesXpath = FALSE,
  titlesExtractMethod = "htmlTitle", divClass = NULL, language = NULL,
  removeString = NULL, encoding = NULL, removeEverythingAfter = NULL,
  removeEverythingBefore = NULL, removeEverythingBeforeDate = NULL)
}
\arguments{
\item{nameOfProject}{Name of 'castarter' project. Must correspond to the name of a folder in the current working directory.}

\item{nameOfWebsite}{Name of a website included in a 'castarter' project. Must correspond to the name of a sub-folder of the project folder.Defaults to NULL. If no nameOfWebsite is provided, exported files are saved in the nameOfProject/Outputs folder.}
}
\value{
A data.frame with the complete dataset.
}
\description{
Extracts metadata and text from local html files.
}
\examples{
dataset <- CreateDatasetFromHtml(nameOfProject, nameOfWebsite)
}

