% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateDictionary.R
\name{CreateDictionary}
\alias{CreateDictionary}
\title{Create dictionaries}
\usage{
CreateDictionary(terms)
}
\arguments{
\item{terms}{A character vector in the format: c("fruit=apple+orange", "animals=cat+dog+cow")}
}
\value{
A dictionary of the 'quanteda' type.
}
\description{
Create dictionaries to be used by castarter functions such as ShowFreq or ShowTS
}
\examples{
CreateDictionary(terms = c("fruit=apple+orange", "animals=cat+dog+cow"))
}

