% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeSeries.R
\name{CreateDistributionTimeSeries}
\alias{CreateDistributionTimeSeries}
\title{Creates a time series graph showing the distribution of documents by date}
\usage{
CreateDistributionTimeSeries(dataset, specificWebsites = NULL,
  rollingAverage = 30, nameOfProject = NULL, nameOfWebsite = NULL)
}
\arguments{
\item{dataset}{A dataset created with 'castarter'.}

\item{specificWebsites}{Character vector indicating which websites (defined by relative nameOfWebsite) have to be included in graph. If left to default, includes all websites present in the dataset.}

\item{rollingAverage}{Integer, defaults to 30. Number of days used to calculate word frequency as shown in the time series. Time series shows word frequency for each date as an average of the N number of days (N=rollingAverage) following the correspondent date.}

\item{nameOfProject}{Name of 'castarter' project. Must correspond to the name of a folder in the current working directory.}

\item{nameOfWebsite}{Name of a website included in a 'castarter' project. Must correspond to the name of a sub-folder of the project folder.}

\item{startDate, }{endDate Character vector with date in the format year-month-date, e.g. "2015-07-14".}

\item{export}{Logical, defaults to FALSE. If TRUE, saves the graph in both png and pdf format. If nameOfProject and nameOfWebsite are provided, in saves the timeseries in the correspondent "Outputs" subfolder.}
}
\value{
A ggplot2 time series showing number of articles published each day.
}
\description{
Creates a time series graph showing the distribution of documents by date.
}
\examples{
CreateDistributionTimeSeries(dataset)
}

