% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanCorpus.R
\name{CreateDtm}
\alias{CreateDtm}
\title{Creates a Document Term Matrix (DTM).}
\usage{
CreateDtm(corpus, removeSparseTerms = NULL)
}
\arguments{
\item{corpus}{A corpus as created by the 'tm' or 'quanteda' package including relevant metadata and typically created with castarter's CreateCorpus() function.}

\item{removeSparseTerms}{A value between 0 and 1, to be passed to the removeSparseTerms function of the 'tm' function.}
}
\value{
A Document Term Matrix (DTM).
}
\description{
Creates a Document Term Matrix (DTM).
}
\examples{
dtm <- CreateDtm(corpus)
}
\keyword{tm}

