% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preliminary.R
\name{CreateFolderStructure}
\alias{CreateFolderStructure}
\title{Creates the folder structure needed by 'castarter'.}
\usage{
CreateFolderStructure(nameOfProject, nameOfWebsite)
}
\description{
It needs to be deleted.
}
\examples{
CreateFolderStructure(nameOfProject, nameOfWebsite)
}
\keyword{toBeRemoved}

