% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/downloadIndex.R
\name{CreateIndexPagesLinks}
\alias{CreateIndexPagesLinks}
\title{Generates links to index pages listing individual articles..}
\usage{
CreateIndexPagesLinks(linkFirstChunkIndex, linkSecondChunkIndex = "",
  firstIndexPage = 1, lastIndexPage = 10, increaseBy = 1,
  dateStyle = "", firstYear = "", lastYear = "", leadingZero = TRUE,
  startDate = "", endDate = "", sortIndexPagesLinks = TRUE,
  dateSeparator = "/", export = FALSE, reversedOrder = FALSE)
}
\arguments{
\item{linkFirstChunkIndex}{First part of index link that does not change in other index pages.}

\item{linkSecondChunkIndex}{Part of index link appneded after the part of the link that varies. If not relevant, may be left empty.}
}
\value{
A character vector of links to index pages.
}
\description{
Generates links to index pages listing individual articles.
}
\examples{
CreateIndexPagesLinks <- CreateIndexPagesLinks("http://www.example.com/news/")
}

