% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createLinks.R
\name{CreateLinks}
\alias{CreateLinks}
\title{Generates links to index pages listing individual articles..}
\usage{
CreateLinks(linkFirstChunk, linkSecondChunk = NULL, startPage = 1,
  endPage = 10, increaseBy = 1, dateFormat = NULL,
  startDate = NULL, endDate = NULL, dateSeparator = "-",
  reversedOrder = FALSE, more = NULL, export = FALSE,
  importParameters = NULL, exportParameters = TRUE, project = NULL,
  website = NULL)
}
\arguments{
\item{linkFirstChunk}{First part of index link that does not change in other index pages.}

\item{linkSecondChunk}{Part of index link appneded after the part of the link that varies. If not relevant, may be left empty.}

\item{startPage}{If the links include a numerical component, define first number of the sequence. startPage defaults to 1.}

\item{endPage}{If the links include a numerical component, define first number of the sequence. endPage defaults to 10.}

\item{increaseBy}{Defines by how much the number in the link should be increased in the numerical sequence. Defaults to 1.}

\item{dateFormat}{A charachter string that defines the format of the date to incldue in the link. Available options are: "Y" (e.g. 2015), "Ym" (2015-10), "Ymd" (e.g. 2015-10-24).}

\item{more}{Defaults to NULL. If given, it can be used to create links based on multiple parameters. FIXME It must be a data frame, create a template with `CreateLinks_template <- tibble::as_tibble(data.frame(matrix(nrow=0,ncol=length(c("linkFirstChunk","linkSecondChunk","startPage", "endPage", "increaseBy","dateFormat","startDate", "endDate",   "dateSeparator", "reversedOrder")))))`; `names(CreateLinks_template) <- params_available` ; `readr::write_csv(x = CreateLinks_template, path = "CreateLinks_template.csv")`}

\item{export}{Exports the links generated by the function in a .txt file in the folder project/website.}

\item{importParameters}{Defaults to NULL. If TRUE, ignores all parameters given in the function call, and imports them from parameters file stored in "project/website/Logs/parameters.rds".}

\item{exportParameters}{Defaults to TRUE. If TRUE, function parameters are exported in the project/website folder. They can be used to update the corpus. Requires parameters project/website.}

\item{project}{Name of 'castarter' project. Must correspond to the name of a folder in the current working directory. Not required if previously set with SetCastarter(project = "project", website = "website")}

\item{website}{Name of a website included in a 'castarter' project. Must correspond to the name of a sub-folder of the project folder. Not required if previously set with SetCastarter(project = "project", website = "website")}
}
\value{
A character vector of links to index pages.
}
\description{
Generates links to index pages listing individual articles.
}
\examples{
\dontrun{
indexLinks <- CreateLinks("http://www.example.com/news/")
}
}
