% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createShinyApp.R
\name{CreateShinyApp}
\alias{CreateShinyApp}
\title{Creates and starts a Shiny App exposing basic 'castarter' functions based on the given dataset}
\usage{
CreateShinyApp(dataset, terms = "europe", customTitle = NULL,
  project = NULL, website = NULL)
}
\arguments{
\item{dataset}{A 'castarter' dataset.}

\item{terms}{String to be shown by default in the newly created app. If a vector of multiple strings is given, they are all included separated by a comma.}

\item{customTitle}{Cutstom title to be given to shiny app.}

\item{project}{Name of 'castarter' project. Must correspond to the name of a folder in the current working directory.}

\item{website}{Name of a website included in a 'castarter' project. Must correspond to the name of a sub-folder of the project folder.}
}
\value{
Nothing, used for side effects: a shiny app is created in the Outputs folder.
}
\description{
Creates and starts a Shiny App exposing basic 'castarter' functions based on the given dataset
}
\examples{
CreateShinyApp(dataset)
}
