% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareCorpus.R
\name{CreateTidyCorpus}
\alias{CreateTidyCorpus}
\title{Converts a 'castarter' dataset into a tidy corpus in line with the 'tidytext' package}
\usage{
CreateTidyCorpus(dataset)
}
\arguments{
\item{dataset}{A data.frame created by 'castarter' including metadata and full text articles to be converted into a corpus.}
}
\value{
A data.frame (tibble) compatible with the 'tidytext' package
}
\description{
Takes a dataset created with 'castarter' and converts it into a tidy corpus of the 'tidytext' type.
}
\examples{
corpus <- CreateCorpus(dataset)
}
