% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/timeSeries.R
\name{CreateTimeSeries}
\alias{CreateTimeSeries}
\title{Creates time series of the frequency of specific terms.}
\usage{
CreateTimeSeries(dataset, corpus, specificTerms, specificWebsites = "",
  startDate = "", rollingAverage = 30)
}
\arguments{
\item{dataset}{A dataset of one or more websites created with 'castarter'.}

\item{corpus}{TM corpus with the same contents of the dataset.}

\item{specificTerms}{Character vector with one or more words to be analysed.}
}
\description{
It creates time series with the frequency of one or more terms, in one or more websites.
}
\examples{
CreateTimeSeries(dataset, corpus, specificTerms = c("word1", "word2"))
}

