% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.R
\name{DivideByWebsite}
\alias{DivideByWebsite}
\title{Divide corpus by website}
\usage{
DivideByWebsite(corpus, nameOfProject)
}
\arguments{
\item{corpus}{A corpus as created by the 'tm' package including metadata.}
}
\value{
A data.frame
}
\description{
Divide corpus by website
}
\examples{
DivideByWebsite(corpus, nameOfProject)
}

