% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadArticles.R
\name{DownloadArticles}
\alias{DownloadArticles}
\title{Downloads files in html format based on list of links.}
\usage{
DownloadArticles(nameOfProject, nameOfWebsite, articlesLinks,
  extractArticlesId = FALSE, start = 1, wait = 1, wget = FALSE)
}
\arguments{
\item{nameOfProject}{Name of 'castarter' project. Must correspond to the name of a folder in the current working directory.}

\item{nameOfWebsite}{Name of a website included in a 'castarter' project. Must correspond to the name of a sub-folder of the project folder.}

\item{start}{ID of articleLinks from which the download should start. Useful for interrupted downloads. Defaults to 1.}
}
\value{
A character vector of html file, and html files saved in dedicated folder.
}
\description{
Downloads files in html format based on list of links. If a vector of articles in html format is provided (typically, articlesHtml), it outputs a character vector including previously and newly downloaded html files. Otherwise, it just downloads html files in the dedicated nameOfProject/nameOfWebsite/Html sub-folder.
}
\examples{
articlesHtml <- DownloadArticles(nameOfProject, nameOfWebsite, articlesLinks)
}

