% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadContents.R
\name{DownloadContents}
\alias{DownloadContents}
\title{Downloads html pages based on a vector of links}
\usage{
DownloadContents(links, type = "articles", nameOfProject, nameOfWebsite,
  articlesHtml = NULL, size = 500, linksToDownload = NULL, wget = FALSE,
  missingArticles = FALSE, wait = 3, createScript = FALSE)
}
\arguments{
\item{links}{A character vector of links, commonly generated either with the function CreateLinks or ExtractLinks.}

\item{type}{Accepted values are either "articles" (default), or "index"; it defines the folder where files are stored.}

\item{nameOfProject}{Name of 'castarter' project. Must correspond to the name of a folder in the current working directory.}

\item{nameOfWebsite}{Name of a website included in a 'castarter' project. Must correspond to the name of a sub-folder of the project folder.}
}
\value{
By default, returns nothing, used for its side effects (downloads html files in relevant folder). Download files can then be imported in a vector with the function ImportHtml.
}
\description{
Downloads html pages based on a vector of links.
}
\examples{
html <- DownloadContents(nameOfProject, nameOfWebsite, links)
}

