% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadContents.R
\name{DownloadContentsForm}
\alias{DownloadContentsForm}
\title{Downloads html pages that are loaded through a web form.}
\usage{
DownloadContentsForm(linkFirstChunk, firstYear = "", lastYear = "",
  dateSeparator = "/", dateBy = "months", start = 1, nameOfProject,
  nameOfWebsite)
}
\arguments{
\item{nameOfProject}{Name of 'castarter' project. Must correspond to the name of a folder in the current working directory.}

\item{nameOfWebsite}{Name of a website included in a 'castarter' project. Must correspond to the name of a sub-folder of the project folder.}
}
\value{
By default, returns nothing, used for its side effects (downloads html files in relevant folder). Download files can then be imported in a vector with the function ImportHtml.
}
\description{
Downloads html pages that are loaded through a web form.
}
\examples{
DownloadContentsForm(links)
}

