% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uploadDownload.R
\name{DownloadDatasets}
\alias{DownloadDatasets}
\title{Downloads a set of datasets from existings 'castarter' projects to an online repository.}
\usage{
DownloadDatasets(projectsAndWebsites, server, user, pwd)
}
\arguments{
\item{projectsAndWebsites}{A character vector listing websites to be loaded in the format "project/website".}
}
\value{
A 'castarter' dataset.
}
\description{
Downloads a set of datasets from existings 'castarter' projects to an online repository.
}
\examples{
projectsAndWebsites <- c("ProjectX/Website1", "ProjectY/Website3", "ProjectZ/Website2")
DownloadDataset(projectsAndWebsites, server = "myserver.com", user = "me@myserver.com", pwd = "secretPassword!")
}
