% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/downloadIndex.R
\name{DownloadIndexPages}
\alias{DownloadIndexPages}
\title{Downloads index pages of a website.}
\usage{
DownloadIndexPages(nameOfProject, nameOfWebsite, indexPagesLinks, start = 1,
  wget = FALSE, phantomjs = FALSE, wait = 0, update = FALSE)
}
\arguments{
\item{nameOfProject}{Name of 'castarter' project. Must correspond to the name of a folder in the current working directory.}

\item{nameOfWebsite}{Name of a website included in a 'castarter' project. Must correspond to the name of a sub-folder of the project folder.}

\item{indexPagesLinks}{A character vector of index pages, possibly generated with the function CreateIndexPagesLinks.}
}
\value{
A character vector of html files.
}
\description{
Downloads index pages of a website.
}
\examples{
indexPagesHtml <- DownloadIndexPages(nameOfProject, nameOfWebsite, indexPagesLinks)
}

