% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cleanCorpus.R
\name{ExportStemmingDictionary}
\alias{ExportStemmingDictionary}
\title{Exports the stemming dictionary}
\usage{
ExportStemmingDictionary(corpus, nameOfProject, stemmingEditMode = "fromCsv")
}
\arguments{
\item{corpus}{A corpus as created by the 'tm' package including metadata.}
}
\value{
.A data.frame with words before and after stemming.
}
\description{
Allows to export a stemming dictionary in csv format.
}
\examples{
stemmingDictionary <- ExportStemmingDictionary(corpus, nameOfProject)
}

