% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractMetadata.R
\name{ExtractArticleId}
\alias{ExtractArticleId}
\title{Extracts articlesId from filename}
\usage{
ExtractArticleId(nameOfProject = NULL, nameOfWebsite = NULL,
  accordingToDate = FALSE, dates = NULL)
}
\arguments{
\item{nameOfProject}{Name of 'castarter' project. Must correspond to the name of a folder in the current working directory.}

\item{nameOfWebsite}{Name of a website included in a 'castarter' project. Must correspond to the name of a sub-folder of the project folder.Defaults to NULL. If no nameOfWebsite is provided, exported files are saved in the nameOfProject/Outputs folder.}
}
\value{
A vector of the integer class.
}
\description{
Extracts articlesId from filename
}
\examples{
articlesId <- ExtractArticleId(nameOfProject, nameOfWebsite)
}

