% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/extractArticlesLinks.R
\name{ExtractArticlesLinks}
\alias{ExtractArticlesLinks}
\title{Extracts direct links to individual articles from index pages.}
\usage{
ExtractArticlesLinks(domain, partOfDirectLink, indexPagesHtml,
  containerType = "", containerClass = "", divClass = "",
  partOfDirectLinkToExclude = "", sort = TRUE, export = FALSE)
}
\arguments{
\item{domain}{Web domain of the website. Will be added at the beginning of each link found.}
}
\value{
A named character vector of links to articles. Name of the link may be the article title.
}
\description{
Extracts direct links to individual articles from index pages according to a selcted pattern.
}
\examples{
articlesLinks <- ExtractArticlesLinks(domain = "http://www.example.com/", partOfDirectLink = "news/", indexPagesHtml)
}

