% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/extractMetadata.R
\name{ExtractDates}
\alias{ExtractDates}
\title{Extracts dates from a vector of html files}
\usage{
ExtractDates(articlesHtml, dateFormat = "dmY", language = "en",
  customString = "")
}
\arguments{
\item{articlesHtml}{A character vector of html files.}

\item{dateFormat}{A string used to extract the date. Available date formats options include dmY, dby, dBy, dBY, dbY, etc.}
}
\value{
A vector of the POSIXct class.
}
\description{
Extracts dates from a vector of html files.
}
\examples{
ExtractDates <- ExtractDates(articlesHtml)
}

