% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/extractMetadata.R
\name{ExtractDatesXpath}
\alias{ExtractDatesXpath}
\title{Extracts dates from a vector of html files}
\usage{
ExtractDatesXpath(articlesHtml, dateFormat = "dmy", divClass = NULL,
  spanClass = "", customXpath = "", language = "en", customString = "",
  minDate = NULL, maxDate = NULL)
}
\arguments{
\item{articlesHtml}{A character vector of html files.}

\item{dateFormat}{A string used to extract the date. Available date formats options include dmY, dby, dBy, dBY, dbY, etc.}

\item{minDate,}{maxDate Minimum and maximum possible dates in the format year-month-date, e.g. "2007-06-24". Introduces NA in the place of impossibly high or low dates.}
}
\value{
A vector of the POSIXct class.
}
\description{
Extracts dates from a vector of html files.
}
\examples{
dates <- ExtractDatesXpath(articlesHtml)
}

