% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractText.R
\name{ExtractText}
\alias{ExtractText}
\title{Extracts textual contents from a vector of html files}
\usage{
ExtractText(container = NULL, containerClass = NULL, containerId = NULL,
  htmlLocation = NULL, id = NULL, metadata = NULL, export = FALSE,
  maxTitleCharacters = 80, removeString = NULL, divClass = NULL,
  divId = NULL, customXpath = NULL, removeEverythingAfter = NULL,
  removeEverythingBefore = NULL, removePunctuationInFilename = TRUE,
  removeTitleFromTxt = FALSE, titles = NULL, keepEverything = FALSE,
  importParameters = NULL, exportParameters = TRUE, progressBar = TRUE,
  project = NULL, website = NULL)
}
\arguments{
\item{htmlLocation}{Path to folder where html files, tipically downloaded with DownloadContents(links) are stored. If not given, it defaults to the Html folder inside project/website folders.}

\item{id}{Defaults to NULL. If provided, it should be a vector of integers. Only html files corresponding to given id in the relevant htmlLocation will be processed.}

\item{metadata}{Defaults to NULL. A data.frame presumably created with ExportMetadata() including information on all articles. Number of rows must correspond to the number of articles to be elaborated. This is required when export == TRUE, in order to provide meaningful filenames.}

\item{export}{Logical, defaults to TRUE. If TRUE, textual contents are saved as individual txt files in a dedicated folder. Filename is based on the medatadata.}

\item{maxTitleCharacters}{Maximum number of characters allowed in the title. Defaults to 80.}

\item{removeString}{A character vector of one or more strings. Provided strings are removed from each article.}

\item{removePunctuationInFilename}{Logical, defaults to TRUE. If TRUE (and export == TRUE), it removes punctuation signs from filemanes to prevent errors in saving files.}

\item{keepEverything}{Logical. If TRUE, the functions calls the boilerpipeR::KeepEverythingExtractor from boilerpipeR, instead of boilerpipeR::ArticleExtractor.}
}
\value{
A character vector of text, and individual articles saved as txt files in a dedicated folder if 'export' is set to TRUE.
}
\description{
Extracts textual contents from a vector of html files.
}
\examples{
text <- ExtractText(container = "div", containerClass = "article")
}
