% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/extractMetadata.R
\name{ExtractTitles}
\alias{ExtractTitles}
\title{Extracts titles of individual pages}
\usage{
ExtractTitles(articlesHtml, articlesLinks = "",
  titlesExtractMethod = "indexLink", removePunctuationInFilename = TRUE,
  remove = "", removeString = "", customXpath = "",
  maximumNumberOfCharactersInTitle = "")
}
\arguments{
\item{articlesHtml}{A character vector of html files.}

\item{articlesLinks}{A named character vector, typically created by the ExtractArticlesLinks function.}
}
\value{
A character vector.
}
\description{
Extracts titles of individual pages from a vector of html files or from a named vector of links.
}
\examples{
titles <- ExtractTitles(articlesHtml)
}

