% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/extractTxt.R
\name{ExtractTxt}
\alias{ExtractTxt}
\title{Extracts textual contents from a vector of html files}
\usage{
ExtractTxt(articlesHtml, metadata = "", export = TRUE,
  maxTitleCharacters = 80, textToBeRemoved = NULL, divClass = NULL,
  removeEverythingAfter = NULL, removeEverythingBefore = NULL,
  removePunctuationInFilename = TRUE, keepEverything = FALSE)
}
\arguments{
\item{articlesHtml}{A character vector of html files.}

\item{export}{Logical, defaults to TRUE. If TRUE, textual contents are saved as individual txt files in a dedicated folder. Filename is based on the medatadata.}

\item{maxTitleCharacters}{Maximum number of characters allowed in the title. Defaults to 80.}

\item{textToBeRemoved}{A character vector of one or more strings. Provided strings are removed from each article.}

\item{divClass}{If provided, all text included in the chosen div is outputted.}

\item{removePunctuationInFilename}{Logical, defaults to TRUE. If TRUE (and export == TRUE), it removes punctuation signs from filemanes to prevent errors in saving files.}

\item{keepEverything}{Logical. If TRUE, the functions calls the boilerpipeR::KeepEverythingExtractor from boilerpipeR, instead of boilerpipeR::ArticleExtractor.}
}
\value{
A character vector of txt files, and individual articles saved as txt files in a dedicated folder if 'export' is set to TRUE.
}
\description{
Extracts textual contents from a vector of html files.
}
\examples{
articlesTxt <- ExtractTxt(articlesHtml, metadata)
}

