% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractMetadata.R
\name{ExtractXpath}
\alias{ExtractXpath}
\title{Extracts string included in an html element}
\usage{
ExtractXpath(articlesHtml, divClass = NULL, spanClass = NULL,
  customXpath = NULL)
}
\arguments{
\item{articlesHtml}{A character vector of html files.}
}
\value{
A character vector with the contents of the given div or custom Xpath.
}
\description{
Extracts string included in an html element
}
\examples{
string <- ExtractDiv(articlesHtml, divClass = "nameOfDiv")
}
