% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareCorpus.R
\name{LoadAllDatasets}
\alias{LoadAllDatasets}
\title{Loads all datasets of a 'castarter' project.}
\usage{
LoadAllDatasets(project, removeNAdates = TRUE, arrangeByDate = TRUE)
}
\arguments{
\item{project}{The name of the project whose datasets are to be imported.}

\item{removeNAdates}{Logical, defaults to TRUE. If TRUE, dataset items that do not have a date in the records are not imported.}

\item{arrangeByDate}{Logical, defaults to TRUE. If TRUE, the dataset is put in order by date (oldest first).}
}
\value{
A data frame including all datasets of a project.
}
\description{
Takes all datasets from all websites in a project and outputs them in a data frame.
}
\examples{
\dontrun{
allDatasets <- LoadAllDatasets(project)
}
}
