% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/prepareCorpus.R
\name{LoadDatasets}
\alias{LoadDatasets}
\title{Loads a specific set of datasets from existings 'castarter' projects.}
\usage{
LoadDatasets(projectsAndWebsites)
}
\arguments{
\item{projectsAndWebsites}{A character vector listing websites to be loaded in the format "nameOfProject/nameOfWebsite".}
}
\value{
A data frame including all loaded datasets.
}
\description{
Takes a specific set of datasets from existings 'castarter' projects.
}
\examples{
projectsAndWebsites <- c("ProjectX/Website1", "ProjectY/Website3", "ProjectZ/Website2")
allDatasets <- LoadDatasets(projectsAndWebsites)
}

