% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractMetadata.R
\name{MergeDates}
\alias{MergeDates}
\title{Merge alternative sets of dates for a given dataset}
\usage{
MergeDates(dates1, dates2, dates3 = "", minDate = "", maxDate = "")
}
\arguments{
\item{dates1, }{dates2, dates3 Vectors of the POSIXct class to be merged. Where a value in dates1 is NA, MergeDates looks for a corresponding value in dates2 and dates3.}

\item{minDate, }{maxDate Minimum and maximum possible dates in the format year-month-date, e.g. "2007-06-24". Introduces NA in the place of impossibly high or low dates.}
}
\value{
A vector of the POSIXct class. .
}
\description{
Merge alternative sets of dates for a given dataset.
}
\examples{
dates <- MergeDates(dates1, dates2)
}

