% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyReadAndTag.R
\name{ReadAndTag}
\alias{ReadAndTag}
\title{Starts a Shiny app that facilitates sub-setting, reading, and tagging a `castarter` dataset.}
\usage{
ReadAndTag()
}
\description{
Starts a Shiny app that facilitates sub-setting, reading, and tagging a `castarter` dataset.
}
\examples{
\dontrun{
ReadAndTag()
}
}
