% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/downloadArticles.R
\name{ReDownloadMissingArticles}
\alias{ReDownloadMissingArticles}
\title{ReDownloads missing articles.}
\usage{
ReDownloadMissingArticles(nameOfProject, nameOfWebsite, links = articlesLinks,
  size = 500, htmlFilesToDownload = "", wget = FALSE, wait = 3)
}
\arguments{
\item{nameOfProject}{Name of 'castarter' project. Must correspond to the name of a folder in the current working directory.}

\item{nameOfWebsite}{Name of a website included in a 'castarter' project. Must correspond to the name of a sub-folder of the project folder.}
}
\description{
Probably needs to be deleted, and integrated into DownloadArticles
}
\examples{
ReDownloadMissingArticles(nameOfProject, nameOfWebsite, articlesLinks)
}

