% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cleanCorpus.R
\name{RemoveStopwords}
\alias{RemoveStopwords}
\title{Remove stopwords from corpus}
\usage{
RemoveStopwords(corpus, stopwords)
}
\arguments{
\item{corpus}{A corpus created with the 'tm' package.}
}
\value{
A corpus without stopwords.
}
\description{
Remove stopwords from corpus.
}
\examples{
corpus <- RemoveStopwords(corpus, stopwords)
}

