% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backup_to_drive.R
\name{restore_from_google_drive}
\alias{restore_from_google_drive}
\title{Restore from Google Drive}
\usage{
restore_from_google_drive(r_files = TRUE, dataset = TRUE,
  archive_files = FALSE, project = NULL, website = NULL)
}
\arguments{
\item{project}{Name of 'castarter' project. Must correspond to the name of a folder in the current working directory. Defaults to NULL, required for storing export parameters (with exportParameters = TRUE). This can be left blank if previously set with SetCastarter(project = "project", website = "website").}

\item{website}{Name of a website included in a 'castarter' project. Must correspond to the name of a sub-folder of the project folder. Defaults to NULL, required for storing export parameters (with exportParameters = TRUE). This can be left blank if previously set with SetCastarter(project = "project", website = "website").}
}
\value{
Nothing, used for its side effects.
}
\description{
Restore from Google Drive
}
\examples{
\dontrun{
restore_from_google_drive()
}
}
