% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archiveFolders.R
\name{RestoreArchives}
\alias{RestoreArchives}
\title{Restore Html, Txt, and IndexHtml folders previously archived with in .tar.gz files with ArchiveFolders().}
\usage{
RestoreArchives(
  html = FALSE,
  indexHtml = FALSE,
  txt = FALSE,
  overwrite = FALSE,
  project = NULL,
  website = NULL
)
}
\arguments{
\item{html}{Logical, defaults to FALSE. If TRUE, restores previously archived html files in their standard folder.}

\item{indexHtml}{Logical, defaults to FALSE. If TRUE, restores previously archived indexHtml files in their standard folder.}

\item{txt}{Logical, defaults to FALSE. If TRUE, restores previously archived txt files in their standard folder.}

\item{project}{Name of 'castarter' project. Must correspond to the name of a folder in the current working directory.}

\item{website}{Name of a website included in a 'castarter' project. Must correspond to the name of a sub-folder of the project folder.}
}
\description{
Restore Html, Txt, and IndexHtml folders previously archived with in .tar.gz files with ArchiveFolders().
}
\section{Warning}{

If the option overwrite is enabled, the function will overwrite files in Html, IndexHtml and Txt folders.
}

\examples{
\dontrun{
RestoreArchives(project, website, removeArchivedFolders = FALSE)
}
}
