% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showMostFrequent.R
\name{ShowMostFrequent}
\alias{ShowMostFrequent}
\title{Shows most frequent terms in a corpus}
\usage{
ShowMostFrequent(corpusDtm, mode = "data.frame", number = 10,
  specificTerms = NULL, stemCompletion = FALSE, corpusOriginal = "",
  minFrequency = 0, export = FALSE, customTitle = NULL, tipology = NULL,
  nameOfProject = NULL, nameOfWebsite = NULL)
}
\arguments{
\item{corpusDtm}{A document term matrix.}

\item{mode}{Defines the type of output. Can be
\itemize{
 \item{"data.frame"}{: Outputs a data.frame This is the default option.}
 \item{"barchart"}{: Outputs a ggplot2 barchart.}
 \item{"wordcloud"}{: Outputs a wordcloud.}
}}

\item{specificTerms}{A character vector, defaults to NULL. If specificTerms is provided, only terms included in this vector will be included in the output.}

\item{export}{Logical, defaults to FALSE. If TRUE, saves the time series in both png and pdf format. If nameOfProject and nameOfWebsite are provided, in saves the barchart in the "Outputs" subfolder.}

\item{corpusDtm}{A document-term matrix or a document-feature matrix of the 'quanteda' type.}

\item{title}{A character vector, defaults to NULL.It allows to customize the title of the exported barchart file.}
}
\value{
A data.frame, barchart, or wordcloud as defined with the 'mode' parameter.
}
\description{
Shows most frequent terms in a corpus starting from a document term matrix.
}
\examples{
corpusDtm <- DocumentTermMatrix(corpus).
mostFrequent <- ShowMostFrequent(corpusDtm)
}

