% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeSeries.R
\name{ShowTS}
\alias{ShowTS}
\title{Creates time series of the frequency of specific terms.}
\usage{
ShowTS(terms, corpusDtm = NULL, corpus = NULL, specificWebsites = NULL,
  startDate = NULL, endDate = NULL, rollingAverage = 30,
  align = "center", verticalLine = NULL, smoothLine = FALSE,
  customTitle = NULL, export = FALSE, allWebsitesAsOne = FALSE,
  dygraphs = FALSE, project = NULL, website = NULL)
}
\arguments{
\item{terms}{A character vector with one or more words to be analysed, or a 'quanteda' dictionary if corpus/corpusDtm are also of the 'quanteda' type.}

\item{corpusDtm}{A document-term matrix or a document-feature matrix of the 'quanteda' type. If provided, other parameters (specificWebsites, startDate, endDate) are ignored, but computation is faster.}

\item{corpus}{A corpus of the 'tm' or 'quanteda' type, presumably created with castarter's ConvertToCorpus() function.}

\item{specificWebsites}{Character vector of the names of one or more websites included in the corpus. Only selected websites will be included in the analysis.}

\item{startDate, }{endDate Character vector with date in the format year-month-date, e.g. "2015-07-14". Given dates are included, e.g. if you wish to include all of 2015, set startDate="2015-01-01", endDate"2015-12-31")}

\item{rollingAverage}{Integer, defaults to 30. Number of days used to calculate word frequency as shown in the time series. Time series shows word frequency for each date as an average of the N number of days (N=rollingAverage) following the correspondent date.}

\item{align}{Defaults to "center", can be either "left", "right" or "center" and refers to the way the rolling average is calculated.}

\item{verticalLine}{Defaults to NULL. Draws a vertical dotted line at the date provided. If given, value must correspond to one or more dates in the YMD format, e.g. "2016-08-17".}

\item{smoothLine}{Logical, defaults to FALSE. If TRUE draws a smooth line over the time series.}

\item{customTitle}{A character vector, defaults to NULL. If provided, it overrides default graph title.}

\item{export}{Logical, defaults to FALSE. If TRUE, saves the time series in both png and pdf format. If project and website are provided, in saves the timeseries in the "Outputs" subfolder.}

\item{project}{Name of 'castarter' project. Must correspond to the name of a folder in the current working directory.}

\item{website}{Name of a website included in a 'castarter' project. Must correspond to the name of a sub-folder of the project folder.}
}
\description{
It creates time series with the frequency of one or more terms, in one or more websites.
}
\examples{
ShowTS(corpus, terms = c("word1", "word2"))
}

