% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinySubsetAndRead.R
\name{SubsetAndRead}
\alias{SubsetAndRead}
\title{Starts a Shiny app that facilitates sub-setting, reading, and tagging a `castarter` dataset.}
\usage{
SubsetAndRead()
}
\description{
Starts a Shiny app that facilitates sub-setting, reading, and tagging a `castarter` dataset.
}
\examples{
\dontrun{
SubsetAndRead()
}
}
