% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{SummariseDataset}
\alias{SummariseDataset}
\title{Summarise key data about a dataset in a data.frame}
\usage{
SummariseDataset(dataset)
}
\arguments{
\item{dataset}{A dataset created with 'castarter'.}
}
\value{
A data.frame with summary statistics.
}
\description{
Summarise key data about a dataset in a data.frame
}
\examples{
SummariseDataset(dataset)
}

