% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UploadDownload.R
\name{UploadDatasets}
\alias{UploadDatasets}
\title{Uploads a set of datasets from existings 'castarter' projects to an online repository.}
\usage{
UploadDatasets(projectsAndWebsites, server, user, pwd)
}
\arguments{
\item{projectsAndWebsites}{A character vector listing websites to be loaded in the format "nameOfProject/nameOfWebsite".}
}
\value{
Nothing. Used for its side effects (uploading datasets to an ftp server).
}
\description{
Uploads a set of datasets from existings 'castarter' projects to an online repository.
}
\examples{
projectsAndWebsites <- c("ProjectX/Website1", "ProjectY/Website3", "ProjectZ/Website2")
UploadDataset(projectsAndWebsites, server = "myserver.com", user = "me@myserver.com", pwd = "secretPassword!")
}

