% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cleanCorpus.R
\name{AddStopwords}
\alias{AddStopwords}
\title{Adds stopwords to the default list}
\usage{
AddStopwords(newStopwords, nameOfProject = NULL, includeDefaultList = FALSE,
  language = "en", importExport = FALSE)
}
\arguments{
\item{newStopwords}{A character vector of words.}

\item{importExport}{Defaults to FALSE. If TRUE, saves stopwords in a txt file in the project folder, and imports changes made in the txt file.}
}
\value{
A character vector.
}
\description{
Adds stopwords to the default list.
}
\examples{
newStopwords <- c("will", "also", "can")
stopwords <- AddStopwords(newStopwords, nameOfProject, includeDefaultList = TRUE, language = "en")
}

