% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanCorpus.R
\name{AddStopwords}
\alias{AddStopwords}
\title{Adds stopwords to the default list}
\usage{
AddStopwords(newStopwords = NULL, project = NULL, includeDefault = TRUE,
  language = "en", importExport = FALSE)
}
\arguments{
\item{newStopwords}{A character vector of words.}

\item{project}{Name of 'castarter' project. Must correspond to the name of a folder in the current working directory.}

\item{includeDefault}{Logical, defaults to TRUE. Includes default list of stopwords included in the 'tm' package.}

\item{language}{Passed to the stopwords function of the TM package to provide a default list of stopwords.}

\item{importExport}{Defaults to FALSE. If TRUE, saves stopwords in a txt file in the project folder, and imports changes made in the txt file.}

\item{importExport}{If TRUE exports the stopword list to a txt file in the project folder. The txt can be edited and re-imported re-running the function.}
}
\value{
A character vector.
}
\description{
Adds stopwords to the default list.
}
\examples{
newStopwords <- c("will", "also", "can")
stopwords <- AddStopwords(newStopwords, project, includeDefault = TRUE, language = "en")
}

