% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cleanCorpus.R
\name{AddWordCombinations}
\alias{AddWordCombinations}
\title{Prepares combinations of words to be changed}
\usage{
AddWordCombinations(wordCombinations = NULL, originalCombination = NULL,
  toBeUsed = NULL, nameOfProject = NULL, dfedit = FALSE,
  importExport = FALSE)
}
\arguments{
\item{originalCombination,}{toBeUsed A character vector.of length 1.}

\item{importExport}{Defaults to FALSE. If TRUE, saves stopwords in a txt file in the project folder, and imports changes made in the txt file.}
}
\value{
.A data.frame of word combinations.
}
\description{
Prepares combinations of words to be changed.
}
\examples{
originalCombination <- "European Union"
toBeUsed <- "europeanunion"
wordCombinations <- AddWordCombinations(originalCombination, toBeUsed, nameOfProject = "")
}

