% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runShinyApps.R
\name{AnalyseDataset}
\alias{AnalyseDataset}
\title{Starts a Shiny app that facilitates sub-setting, reading, and tagging a `castarter` dataset.}
\usage{
AnalyseDataset()
}
\description{
Starts a Shiny app that facilitates sub-setting, reading, and tagging a `castarter` dataset.
}
\examples{
AnalyseDataset()
}
