% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/analyseTerms.R
\name{AnalyseTerms}
\alias{AnalyseTerms}
\title{Compares frequency of specific words across websites.}
\usage{
AnalyseTerms(corpus, nameOfProject, terms, mode = "graph",
  includeOnly = NULL, order = "", tipology = "", frequency = "relative")
}
\arguments{
\item{corpus}{A TM corpus.}

\item{nameOfProject}{Name of 'castarter' project. Must correspond to the name of a folder in the current working directory.}

\item{terms}{Character vector of terms to be compared.}
}
\description{
Compares frequency of specific words across websites.
}
\examples{
AnalyseTerms(corpus, nameOfProject, terms)
}

