% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/prepareCorpus.R
\name{ConvertToCorpus}
\alias{ConvertToCorpus}
\title{Converts a 'castarter' dataset into a 'tm' corpus.}
\usage{
ConvertToCorpus(dataset)
}
\arguments{
\item{dataset}{A data.frame created by 'castarter' including metadata and full text articles to be converted into a corpus.}
}
\value{
A corpus as created by the 'tm' package including metadata.
}
\description{
Takes a dataset created with 'castarter' and converts it into a 'tm' corpus. Metadata are automatically imported.
}
\examples{
corpus <- ConvertToCorpus(dataset)
}
\keyword{tm}

