% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runShinyApps.R
\name{CreateDataset}
\alias{CreateDataset}
\title{Starts a Shiny app that facilitates the creation of a 'castarter' dataset.}
\usage{
CreateDataset()
}
\description{
Starts a Shiny app that facilitates the creation of a 'castarter' dataset.
}
\examples{
CreateDataset()
}
