% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractMetadata.R
\name{CreateDatasetFromHtml}
\alias{CreateDatasetFromHtml}
\title{Extracts metadata and text from local html files}
\usage{
CreateDatasetFromHtml(articlesLinks = NULL, dateFormat = NULL,
  divClass_ExtractDates = NULL, spanClass_ExtractDates = NULL,
  customXpath_ExtractDates = NULL,
  language_ExtractDates = Sys.getlocale(category = "LC_TIME"),
  removeEverythingBefore_ExtractDates = NULL,
  method_ExtractTitles = "htmlTitle", removeString_ExtractTitles = NULL,
  removeEverythingAfter_ExtractTitles = NULL,
  removePunctuation_ExtractTitles = NULL, divClass_ExtractTxt = NULL,
  divId_ExtractTxt = NULL, removeString_ExtractTxt = NULL,
  removeEverythingAfter_ExtractTxt = NULL,
  removeEverythingBefore_ExtractTxt = NULL, removeTitleFromTxt = FALSE,
  customXpath_ExtractTxt = NULL, language = "English", encoding = NULL,
  logProgress = FALSE, exportParameters = TRUE, nameOfProject = NULL,
  nameOfWebsite = NULL)
}
\arguments{
\item{nameOfProject}{Name of 'castarter' project. Must correspond to the name of a folder in the current working directory.}

\item{nameOfWebsite}{Name of a website included in a 'castarter' project. Must correspond to the name of a sub-folder of the project folder.Defaults to NULL. If no nameOfWebsite is provided, exported files are saved in the nameOfProject/Outputs folder.}
}
\value{
A data.frame with the complete dataset.
}
\description{
Extracts metadata and text from local html files.
}
\examples{
dataset <- CreateDatasetFromHtml()
}

