% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preliminary.R
\name{CreateFolders}
\alias{CreateFolders}
\title{Creates the folder structure needed by 'castarter'.}
\usage{
CreateFolders(project = NULL, website = NULL)
}
\description{
Creates all the standard folders required by a 'castarter' project.
}
\examples{
CreateFolders(project, website)
}
