% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preliminary.R
\name{CreateFolderStructure}
\alias{CreateFolderStructure}
\title{Creates the folder structure needed by 'castarter'.}
\usage{
CreateFolderStructure(nameOfProject, nameOfWebsite)
}
\description{
Creates all the standard folders required by a 'castarter' project.
}
\examples{
CreateFolderStructure(nameOfProject, nameOfWebsite)
}

