% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/downloadIndex.R
\name{CreateLinks}
\alias{CreateLinks}
\title{Generates links to index pages listing individual articles..}
\usage{
CreateLinks(linkFirstChunk, linkSecondChunk = NULL, startPage = 1,
  endPage = 10, increaseBy = 1, dateStyle = "", firstYear = "",
  lastYear = "", leadingZero = TRUE, startDate = "", endDate = "",
  sortindexLinks = TRUE, dateSeparator = "/", export = FALSE,
  reversedOrder = FALSE, exportParameters = FALSE, nameOfProject = NULL,
  nameOfWebsite = NULL)
}
\arguments{
\item{linkFirstChunk}{First part of index link that does not change in other index pages.}

\item{linkSecondChunk}{Part of index link appneded after the part of the link that varies. If not relevant, may be left empty.}

\item{startPage,}{endPage If the links include a numerical component, define first and last number of the sequence.startPage defaults to 1, endPage defaults to 10.}

\item{increaseBy}{Defines by how much the number in the link should be increased in the numerical sequence. Defaults to 1.}
}
\value{
A character vector of links to index pages.
}
\description{
Generates links to index pages listing individual articles.
}
\examples{
indexLinks <- CreateLinks("http://www.example.com/news/")
}

