% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadIndex.R
\name{CreateLinks}
\alias{CreateLinks}
\title{Generates links to index pages listing individual articles..}
\usage{
CreateLinks(linkFirstChunk, linkSecondChunk = NULL, startPage = 1,
  endPage = 10, increaseBy = 1, dateFormat = NULL, firstYear = "",
  lastYear = "", leadingZero = TRUE, startDate = "", endDate = "",
  sortIndexLinks = TRUE, dateSeparator = "/", export = FALSE,
  reversedOrder = FALSE, exportParameters = FALSE, nameOfProject = NULL,
  nameOfWebsite = NULL)
}
\arguments{
\item{linkFirstChunk}{First part of index link that does not change in other index pages.}

\item{linkSecondChunk}{Part of index link appneded after the part of the link that varies. If not relevant, may be left empty.}

\item{startPage, }{endPage If the links include a numerical component, define first and last number of the sequence.startPage defaults to 1, endPage defaults to 10.}

\item{increaseBy}{Defines by how much the number in the link should be increased in the numerical sequence. Defaults to 1.}

\item{dateFormat}{A charachter string that defines the format of the date to incldue in the link. Available options are: "Y" (e.g. 2015), "Ym" (2015-10), "Ymd" (e.g. 2015-10-24).}

\item{leadingZero}{Defaults to TRUE. If TRUE, days and months of one digit are preceded by a zero (e.g. 07/05/2014).}

\item{export}{Exports the links in a .txt file in the folder nameOfProject/nameOfWebsite.}

\item{exportParameters}{Defaults to FALSE. If TRUE, function parameters are exported in the nameOfProject/nameOfWebsite folder. They can be used to update the corpus. Requires parameters nameOfProject/nameOfWebsite.}

\item{nameOfProject}{Name of 'castarter' project. Must correspond to the name of a folder in the current working directory.}
}
\value{
A character vector of links to index pages.
}
\description{
Generates links to index pages listing individual articles.
}
\examples{
indexLinks <- CreateLinks("http://www.example.com/news/")
}

