% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createShinyApp.R
\name{CreateShinyApp}
\alias{CreateShinyApp}
\title{Creates and starts a Shiny App exposing basic 'castarter' functions based on the given dataset}
\usage{
CreateShinyApp(dataset, terms = "europ*", nameOfProject = NULL,
  nameOfWebsite = NULL)
}
\arguments{
\item{dataset}{A 'castarter' dataset or a corpus created with 'castarter' of the quanteda type.}

\item{nameOfProject}{Name of 'castarter' project. Must correspond to the name of a folder in the current working directory.}

\item{nameOfWebsite}{Name of a website included in a 'castarter' project. Must correspond to the name of a sub-folder of the project folder.}
}
\value{
Nothing, used for side effects: a shiny app is created in the Outputs folder.
}
\description{
Creates and starts a Shiny App exposing basic 'castarter' functions based on the given dataset
}
\examples{
CreateShinyApp(dataset)
}

