% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/timeSeries.R
\name{CreateTimeSeries}
\alias{CreateTimeSeries}
\title{Creates time series of the frequency of specific terms.}
\usage{
CreateTimeSeries(corpus, specificTerms, specificWebsites = "",
  startDate = NULL, endDate = NULL, rollingAverage = 30)
}
\arguments{
\item{specificTerms}{Character vector with one or more words to be analysed.}

\item{startDate,}{endDate Character vector with date in the format year-month-date, e.g. "2015-07-14".}

\item{corpusDtm}{A document term matrix.}
}
\description{
It creates time series with the frequency of one or more terms, in one or more websites.
}
\examples{
CreateTimeSeries(corpus, specificTerms = c("word1", "word2"))
}

