% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/downloadArticles.R
\name{DownloadArticles}
\alias{DownloadArticles}
\title{Downloads files in html format based on list of links.}
\usage{
DownloadArticles(nameOfProject, nameOfWebsite, articlesLinks,
  extractArticlesId = FALSE, start = 1, wait = 1, wget = FALSE)
}
\arguments{
\item{nameOfProject}{Name of 'castarter' project. Must correspond to the name of a folder in the current working directory.}

\item{nameOfWebsite}{Name of a website included in a 'castarter' project. Must correspond to the name of a sub-folder of the project folder.}
}
\value{
A character vector of html file, and html files saved in dedicated folder.
}
\description{
Downloads files in html format based on list of links.
}
\examples{
articlesHtml <- DownloadArticles(nameOfProject, nameOfWebsite, articlesLinks)
}

