% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadContents.R
\name{DownloadContents}
\alias{DownloadContents}
\title{Downloads html pages based on a vector of links}
\usage{
DownloadContents(links, type = "articles", articlesHtml = NULL,
  size = 500, linksToDownload = NULL, wgetSystem = FALSE,
  method = "auto", missingArticles = TRUE, start = 1, wait = 1,
  createScript = FALSE, nameOfProject = NULL, nameOfWebsite = NULL)
}
\arguments{
\item{links}{A character vector of links, commonly generated either with the function CreateLinks or ExtractLinks.}

\item{type}{Accepted values are either "articles" (default), or "index"; it defines the folder where files are stored.}

\item{linksToDownload}{A numeric vector. Only corresponding links will be downloaded: links[linksToDownload]}

\item{wgetSystem}{Logical, defaults to FALSE. Calls wget as a system command through the system() function. Wget must be previously installed on the system.}

\item{method}{Defaults to "auto". Method is passed to the function utils::download.file(); available options are "internal", "wininet" (Windows only) "libcurl", "wget" and "curl". For more information see ?utils::download.file()}

\item{missingArticles}{Logical, defaults to TRUE. If TRUE, verifies if a downloaded html file exists for each element in articlesLinks; when there is no such file, it downloads it.}

\item{start}{Integer. Only links with position higher than start in the links vector will be downloaded: links[start:length(links)]}

\item{nameOfProject}{Name of 'castarter' project. Must correspond to the name of a folder in the current working directory.}

\item{nameOfWebsite}{Name of a website included in a 'castarter' project. Must correspond to the name of a sub-folder of the project folder.}
}
\value{
By default, returns nothing, used for its side effects (downloads html files in relevant folder). Download files can then be imported in a vector with the function ImportHtml.
}
\description{
Downloads html pages based on a vector of links.
}
\examples{
DownloadContents(nameOfProject, nameOfWebsite, links)
}

