% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{ExportArticlesWith}
\alias{ExportArticlesWith}
\title{Exports all articles that contain a given term.}
\usage{
ExportArticlesWith(dataset, term, nameOfProject, nameOfWebsite = NULL,
  txt = TRUE, csv = FALSE, xlsx = FALSE, data.frame = FALSE,
  includeOnly = NULL, sortBy = "date")
}
\arguments{
\item{dataset}{A dataset created with 'castarter'.}

\item{term}{The term that determines which articles are exported.Must be a character vector of length = 1.}

\item{nameOfProject}{Name of 'castarter' project. Must correspond to the name of a folder in the current working directory.}

\item{nameOfWebsite}{Name of a website included in a 'castarter' project. Must correspond to the name of a sub-folder of the project folder.Defaults to NULL. If no nameOfWebsite is provided, exported files are saved in the nameOfProject/Outputs folder.}

\item{txt}{Logical, defaults to TRUE. If TRUE, exports all articles including the term provided in a .txt file.}

\item{csv}{Logical, defaults to FALSE. If TRUE, exports all articles including the term provided in a .csv file.}

\item{xlsx}{Logical, defaults to FALSE. If TRUE, exports all articles including the term provided in a .xlsx file.}

\item{data.fram}{Logical, defaults to FALSE. If TRUE, the function outputs a data frame. If FALSE, function used only for its side effects (i.e. exporting to files)}
}
\description{
Exports all articles that contain a given term.
}
\examples{
ExportArticlesWith(dataset, "example", nameOfProject, nameOfWebsite)
}

