% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{ExportArticlesWith}
\alias{ExportArticlesWith}
\title{Exports all articles that contain a given term.}
\usage{
ExportArticlesWith(dataset, term, includeNaDates = TRUE,
  onlyNaDates = FALSE, txt = TRUE, csv = FALSE, xlsx = FALSE,
  data.frame = FALSE, includeOnly = NULL, sortBy = "date",
  project = NULL, website = NULL)
}
\arguments{
\item{dataset}{A dataset created with 'castarter'.}

\item{term}{The term that determines which articles are exported.Must be a character vector of length = 1.}

\item{onlyNaDates}{Logical, defaults to FALSE. Used to for troubleshooting dataset.}

\item{txt}{Logical, defaults to TRUE. If TRUE, exports all articles including the term provided in a .txt file.}

\item{csv}{Logical, defaults to FALSE. If TRUE, exports all articles including the term provided in a .csv file.}

\item{xlsx}{Logical, defaults to FALSE. If TRUE, exports all articles including the term provided in a .xlsx file.}

\item{data.frame}{Logical, defaults to FALSE. If TRUE, the function outputs a data frame. If FALSE, function used only for its side effects (i.e. exporting to files)}

\item{project}{Name of 'castarter' project. Must correspond to the name of a folder in the current working directory.}

\item{website}{Name of a website included in a 'castarter' project. Must correspond to the name of a sub-folder of the project folder.Defaults to NULL. If no website is provided, exported files are saved in the project/Outputs folder.}
}
\description{
Exports all articles that contain a given term.
}
\examples{
ExportArticlesWith(dataset, "example")
}

