% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractMetadata.R
\name{ExportDataset}
\alias{ExportDataset}
\title{Exports dataset}
\usage{
ExportDataset(
  dataset = NULL,
  text = NULL,
  metadata = NULL,
  exportRds = FALSE,
  exportRdata = FALSE,
  exportCsv = FALSE,
  exportXlsx = FALSE,
  project = NULL,
  website = NULL
)
}
\arguments{
\item{dataset}{A 'castarter' dataset.}

\item{text}{A character vector typically created with ExtractText().}

\item{metadata}{A data.frame typically created with ExportMetadata(). Number of rows must be the same as the number of text items in `text`.}

\item{exportCsv}{If equal to TRUE, exports the complete dataset in the .csv file format in the Dataset sub-folder.}

\item{exportXlsx}{If equal to TRUE, exports the complete dataset in the .xlsx file format in the Dataset sub-folder.}

\item{project}{Name of 'castarter' project. Must correspond to the name of a folder in the current working directory.}

\item{website}{Name of a website included in a 'castarter' project. Must correspond to the name of a sub-folder of the project folder.}

\item{exportRdsIf}{equal to TRUE, exports the complete dataset in R's .rds file format in the Dataset sub-folder.}
}
\value{
A data.frame, a 'castarter' dataset.
}
\description{
Exports dataset to a data.frame, with options to save it as an R object, and export it as a .csv or .xlsx file.
}
\examples{
\dontrun{
dataset <- ExportDataset(text, metadata, project, website)
}
}
