% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/extractMetadata.R
\name{ExportMetadata}
\alias{ExportMetadata}
\title{Exports metadata}
\usage{
ExportMetadata(nameOfProject, nameOfWebsite, dates, articlesId, titles,
  language, articlesLinks, exportXlsx = FALSE, accordingToDate = FALSE,
  ignoreNAdates = FALSE, onlyExistingHtmlFiles = FALSE)
}
\arguments{
\item{nameOfProject}{Name of 'castarter' project. Must correspond to the name of a folder in the current working directory.}

\item{nameOfWebsite}{Name of a website included in a 'castarter' project. Must correspond to the name of a sub-folder of the project folder.}

\item{exportXlsx}{If equal to TRUE, exports the complete dataset in the .xlsx file format in the Dataset sub-folder.}
}
\value{
A vector of the POSIXct class.
}
\description{
Exports metadata to a csv file.
}
\examples{
metadata <- ExportMetadata(nameOfProject, nameOfWebsite, dates, articlesId, titles, language, articlesLinks)
}

