% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractMetadata.R
\name{ExportMetadata}
\alias{ExportMetadata}
\title{Exports metadata}
\usage{
ExportMetadata(dates, id, titles, language, links, exportXlsx = FALSE,
  accordingToDate = FALSE, ignoreNAdates = FALSE,
  onlyExistingHtmlFiles = FALSE, project = NULL, website = NULL)
}
\arguments{
\item{exportXlsx}{If equal to TRUE, exports the complete dataset in the .xlsx file format in the Dataset sub-folder.}

\item{project}{Name of 'castarter' project. Must correspond to the name of a folder in the current working directory.}

\item{website}{Name of a website included in a 'castarter' project. Must correspond to the name of a sub-folder of the project folder.}
}
\value{
A vector of the data.frame class.
}
\description{
Exports metadata to a vector, csv or xlsx file.
}
\examples{
metadata <- ExportMetadata(project, website, dates, id, titles, language, links)
}

