% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanCorpus.R
\name{ExportStemmingDictionary}
\alias{ExportStemmingDictionary}
\title{Exports the stemming dictionary}
\usage{
ExportStemmingDictionary(corpusDtm, nameOfProject = NULL, stopwords = "",
  language = "english")
}
\arguments{
\item{corpusDtm}{A document-term matrix.}

\item{nameOfProject}{Name of 'castarter' project. Must correspond to the name of a folder in the current working directory.}

\item{stopwords}{A character vector of stopwords.}

\item{nameOfWebsite}{Name of a website included in a 'castarter' project. Must correspond to the name of a sub-folder of the project folder.}
}
\value{
.A data.frame with words before and after stemming.
}
\description{
Allows to export a stemming dictionary in csv format.
}
\examples{
stemmingDictionary <- ExportStemmingDictionary(corpus, nameOfProject)
}

