% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractMetadata.R
\name{ExtractDates}
\alias{ExtractDates}
\title{Extracts dates from a vector of html files}
\usage{
ExtractDates(articlesHtml, dateFormat = "dmY", language = "en",
  customString = "", minDate = NULL, maxDate = NULL,
  removeEverythingBefore = NULL, keepAllString = FALSE,
  exportParameters = TRUE, nameOfProject = NULL, nameOfWebsite = NULL)
}
\arguments{
\item{articlesHtml}{A character vector of html files.}

\item{dateFormat}{A string used to extract the date. Available date formats options include dmY, dby, dBy, dBY, dbY, etc.}

\item{minDate, }{maxDate Minimum and maximum possible dates in the format year-month-date, e.g. "2007-06-24". Introduces NA in the place of impossibly high or low dates.}

\item{keepAllString}{Logical, defaults to FALSE. If TRUE, it directly tries to parse the date with the given dateFormat, without trying to polish the string provided accordingly.}

\item{nameOfProject}{Name of 'castarter' project. Must correspond to the name of a folder in the current working directory. Required for storing export parameters (with exportParameters = TRUE).}
}
\value{
A vector of the POSIXct class.
}
\description{
Extracts dates from a vector of html files.
}
\examples{
dates <- ExtractDates(articlesHtml)
}

