% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractMetadata.R
\name{ExtractDates}
\alias{ExtractDates}
\title{Extracts dates from a vector of html files}
\usage{
ExtractDates(articlesHtml, dateFormat = "dmY", divClass = NULL,
  divId = NULL, spanClass = NULL, customXpath = NULL,
  language = Sys.getlocale(category = "LC_TIME"), customString = "",
  minDate = NULL, maxDate = NULL, encoding = "UTF-8",
  keepAllString = FALSE, removeEverythingBefore = NULL,
  exportParameters = TRUE, project = NULL, website = NULL)
}
\arguments{
\item{articlesHtml}{A character vector of html files.}

\item{dateFormat}{A string expressing the date format. In line with standards (see ?strptime), 'd' stands for day, 'm' stands for month in figures, 'b' for months spelled out as words, 'y' as year without the century, 'Y' as year with four digits. Standard separation marks among parts of the date (e.g. '-', '/', '.') should not be included. The following date formats are available :
\itemize{
 \item{"dmY"}{: Default.}
 \item{"Ymd"}{: }
 \item{"dbY"}{: }
 \item{"YBd"}{: }
 \item{"dB,Y"}{: }
 \item{"db.'y"}{: }
 \item{"Bd,Y"}{: }
 \item{"xdBY"}{: customString must be provided.}
}}

\item{language}{Provide a language in order to extract name of months. Defaults to the locale currently active in R (usually, the system language). Generic forms such as "english" or "russian", are usually accepted. See ?locales for more details. On linux, you can run system("locale -a", intern = TRUE) to see all available locales.}

\item{minDate, }{maxDate Minimum and maximum possible dates in the format year-month-date, e.g. "2007-06-24". Introduces NA in the place of impossibly high or low dates.}

\item{encoding}{Defaults to NULL. If source is not in UTF, encoding can be specified here for conversion. A list of valid values can be found using iconvlist().}

\item{keepAllString}{Logical, defaults to FALSE. If TRUE, it directly tries to parse the date with the given dateFormat, without trying to polish the string provided accordingly.}

\item{project}{Name of 'castarter' project. Must correspond to the name of a folder in the current working directory. Defaults to NULL, required for storing export parameters (with exportParameters = TRUE). This can be left blank if previously set with SetCastarter(project = "project", website = "website").}

\item{website}{Name of a website included in a 'castarter' project. Must correspond to the name of a sub-folder of the project folder. Defaults to NULL, required for storing export parameters (with exportParameters = TRUE). This can be left blank if previously set with SetCastarter(project = "project", website = "website").}
}
\value{
A vector of the POSIXct class.
}
\description{
Extracts dates from a vector of html files.
}
\examples{
dates <- ExtractDates(articlesHtml)
}
