% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractMetadata.R
\name{ExtractDatesXpath}
\alias{ExtractDatesXpath}
\title{Extracts dates from a vector of html files}
\usage{
ExtractDatesXpath(articlesHtml, dateFormat = "dmy", divClass = NULL,
  spanClass = NULL, customXpath = NULL, language = "english",
  customString = "", minDate = NULL, maxDate = NULL, encoding = NULL,
  keepAllString = FALSE, exportParameters = FALSE, nameOfProject = NULL,
  nameOfWebsite = NULL)
}
\arguments{
\item{articlesHtml}{A character vector of html files.}

\item{dateFormat}{A string used to extract the date. Available date formats options include dmY, dby, dBy, dBY, dbY, etc.}

\item{minDate, }{maxDate Minimum and maximum possible dates in the format year-month-date, e.g. "2007-06-24". Introduces NA in the place of impossibly high or low dates.}

\item{encoding}{Defaults to NULL. If source is not in UTF, encoding can be specified here for conversion. A list of valid values can be found using iconvlist().}

\item{keepAllString}{Logical, defaults to FALSE. If TRUE, it directly tries to parse the date with the given dateFormat, without trying to polish the string provided accordingly.}

\item{nameOfProject}{Name of 'castarter' project. Must correspond to the name of a folder in the current working directory. Required for storing export parameters (with exportParameters = TRUE).}
}
\value{
A vector of the POSIXct class.
}
\description{
Extracts dates from a vector of html files.
}
\examples{
dates <- ExtractDatesXpath(articlesHtml)
}

