% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractMetadata.R
\name{ExtractId}
\alias{ExtractId}
\title{Extracts id from filename}
\usage{
ExtractId(project = NULL, website = NULL, accordingToDate = FALSE,
  dates = NULL)
}
\arguments{
\item{project}{Name of 'castarter' project. Must correspond to the name of a folder in the current working directory.}

\item{website}{Name of a website included in a 'castarter' project. Must correspond to the name of a sub-folder of the project folder.Defaults to NULL. If no website is provided, exported files are saved in the project/Outputs folder.}
}
\value{
A vector of the integer class.
}
\description{
Extracts id from filename
}
\examples{
id <- ExtractId(project, website)
}

