% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractLinks.R
\name{ExtractLinks}
\alias{ExtractLinks}
\title{Extracts direct links to individual articles from index pages.}
\usage{
ExtractLinks(domain, partOfLink, html, containerType = NULL,
  containerClass = NULL, divClass = NULL, attributeType = NULL,
  partOfLinkToExclude = NULL, minLength = NULL, maxLength = NULL,
  indexLinks = NULL, sortLinks = TRUE, linkTitle = TRUE, export = FALSE,
  appendString = NULL, removeString = NULL, progressBar = TRUE,
  exportParameters = TRUE, project = NULL, website = NULL)
}
\arguments{
\item{domain}{Web domain of the website. Will be added at the beginning of each link found.If links in the page already include the full web address this should be ignored. Defaults to "".}

\item{partOfLink}{Part of URL found only in links of individual articles to be downloaded. If more than one provided, it includes all links that contains either of the strings provided.}

\item{containerType}{Type of html container from where links are to be extracted, such as "ul", "div", and others. containerClass must also be provided.}

\item{attributeType}{Type of attribute to extract from links, when different from href.}

\item{partOfLinkToExclude}{If an URL includes this string, it is excluded from the output. One or more strings may be provided.}

\item{minLength}{If a link is shorter than the number of characters given in minLength, it is excluded from the output.}

\item{maxLength}{If a link is longer than the number of characters given in maxLength, it is excluded from the output.}

\item{indexLinks}{A character vector, defaults to NULL. If provided, indexLinks are removed from the extracted articlesLinks.}

\item{sortLinks}{Defaults to TRUE. If TRUE, links are sorted in aphabetical order.}

\item{linkTitle}{Defaults to TRUE. If TRUE, text of links is included as names of the vector.}

\item{appendString}{If provided, appends given string to the extracted articles. Typically used to create links for print or mobile versions of the extracted page.}

\item{removeString}{If provided, remove given string (or strings) from links.}

\item{progressBar}{Logical, defaults to TRUE. If FALSE, progress bar is not shown (useful for example when including scripts in rmarkdown)}

\item{exportParameters}{Defaults to FALSE. If TRUE, function parameters are exported in the project/website folder. They can be used to update the corpus.}
}
\value{
A named character vector of links to articles. Name of the link may be the article title.
}
\description{
Extracts direct links to individual articles from index pages according to a selcted pattern.
}
\examples{
articlesLinks <- ExtractLinks(domain = "http://www.example.com/", partOfLink = "news/", html)
}
