% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/extractArticlesLinks.R
\name{ExtractLinks}
\alias{ExtractLinks}
\title{Extracts direct links to individual articles from index pages.}
\usage{
ExtractLinks(domain, partOfLink, indexPagesHtml, containerType = "",
  containerClass = "", divClass = "", partOfLinkToExclude = "",
  sort = TRUE, export = FALSE)
}
\arguments{
\item{domain}{Web domain of the website. Will be added at the beginning of each link found.If links in the page already include the full web address this should be ignored. Defaults to "".}
}
\value{
A named character vector of links to articles. Name of the link may be the article title.
}
\description{
Extracts direct links to individual articles from index pages according to a selcted pattern.
}
\examples{
articlesLinks <- ExtractLinks(domain = "http://www.example.com/", partOfLink = "news/", indexPagesHtml)
}

