% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractArticlesLinks.R
\name{ExtractLinks}
\alias{ExtractLinks}
\title{Extracts direct links to individual articles from index pages.}
\usage{
ExtractLinks(domain, partOfLink, indexHtml, containerType = "",
  containerClass = "", divClass = "", partOfLinkToExclude = NULL,
  minLength = NULL, indexLinks = NULL, sortLinks = TRUE, export = FALSE,
  appendString = NULL, exportParameters = FALSE, nameOfProject = NULL,
  nameOfWebsite = NULL)
}
\arguments{
\item{domain}{Web domain of the website. Will be added at the beginning of each link found.If links in the page already include the full web address this should be ignored. Defaults to "".}

\item{partOfLink}{Part of URL found only in links of individual articles to be downloaded.}

\item{partOfLinkToExclude}{If an URL includes this string, it is excluded from the output. One or more strings may be provided.}

\item{minLength}{If a link is shorter than the number of characters given in minLength, it is excluded from the output.}

\item{indexLinks}{A character vector, defaults to NULL. If provided, indexLinks are removed from the extracted articlesLinks.}

\item{appendString}{If provided, appends given string to the extracted articles. Typically used to create links for print or mobile versions of the extracted page.}

\item{exportParameters}{Defaults to FALSE. If TRUE, function parameters are exported in the nameOfProject/nameOfWebsite folder. They can be used to update the corpus.}
}
\value{
A named character vector of links to articles. Name of the link may be the article title.
}
\description{
Extracts direct links to individual articles from index pages according to a selcted pattern.
}
\examples{
articlesLinks <- ExtractLinks(domain = "http://www.example.com/", partOfLink = "news/", indexHtml)
}

