% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/extractTxt.R
\name{ExtractTxt}
\alias{ExtractTxt}
\title{Extracts textual contents from a vector of html files}
\usage{
ExtractTxt(articlesHtml, metadata, numberOfTitleCharacters = 80,
  textToBeRemoved = "", removeEverythingAfter = "",
  removePunctuationInFilename = TRUE, keepEverything = FALSE)
}
\arguments{
\item{articlesHtml}{A character vector of html files.}

\item{keepEverything}{Logical. If TRUE, the functions calls the KeepEverythingExtractor from boilerpipeR, instead of ArticleExtractor.}
}
\value{
A character vector of txt files, and individual articles saves ad txt files in a dedicated folder.
}
\description{
Extracts textual contents from a vector of html files.
}
\examples{
articlesTxt <- ExtractTxt(articlesHtml, metadata)
}

