% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractTxt.R
\name{ExtractTxt}
\alias{ExtractTxt}
\title{Extracts textual contents from a vector of html files}
\usage{
ExtractTxt(articlesHtml, metadata = NULL, export = FALSE,
  maxTitleCharacters = 80, stringToBeRemoved = NULL, divClass = NULL,
  divId = NULL, removeEverythingAfter = NULL,
  removeEverythingBefore = NULL, removePunctuationInFilename = TRUE,
  removeTitleFromTxt = FALSE, titles = NULL, keepEverything = FALSE,
  exportParameters = TRUE, nameOfProject = NULL, nameOfWebsite = NULL)
}
\arguments{
\item{articlesHtml}{A character vector of html files.}

\item{metadata}{Defaults to NULL. A data.frame presumably created with ExportMetadata() including information on all articles. Number of rows must correspond to the number of articles to be elaborated. This is required when export == TRUE, in order to provide meaningful filenames.}

\item{export}{Logical, defaults to TRUE. If TRUE, textual contents are saved as individual txt files in a dedicated folder. Filename is based on the medatadata.}

\item{maxTitleCharacters}{Maximum number of characters allowed in the title. Defaults to 80.}

\item{stringToBeRemoved}{A character vector of one or more strings. Provided strings are removed from each article.}

\item{divClass}{If provided, all text included in the chosen div is outputted.}

\item{removePunctuationInFilename}{Logical, defaults to TRUE. If TRUE (and export == TRUE), it removes punctuation signs from filemanes to prevent errors in saving files.}

\item{keepEverything}{Logical. If TRUE, the functions calls the boilerpipeR::KeepEverythingExtractor from boilerpipeR, instead of boilerpipeR::ArticleExtractor.}
}
\value{
A character vector of txt files, and individual articles saved as txt files in a dedicated folder if 'export' is set to TRUE.
}
\description{
Extracts textual contents from a vector of html files.
}
\examples{
articlesTxt <- ExtractTxt(articlesHtml, metadata)
}

