% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/importHtml.R
\name{ImportHtml}
\alias{ImportHtml}
\title{Imports html files into a character vector.}
\usage{
ImportHtml(from = "", nameOfProject = "", nameOfWebsite = "",
  pathToHtmlFolder = "")
}
\arguments{
\item{from}{Allows to choose which type of contents to import. Can be either "articles" or "index".}

\item{nameOfProject}{Name of 'castarter' project. Must correspond to the name of a folder in the current working directory.}

\item{nameOfWebsite}{Name of a website included in a 'castarter' project. Must correspond to the name of a sub-folder of the project folder.}
}
\value{
A character vector of html files.
}
\description{
Imports html files already downloaded into a character vector.
}
\examples{
articlesHtml <- ImportHtml(from = articles, nameOfProject, nameOfWebsite)
}

