% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importHtml.R
\name{ImportHtml}
\alias{ImportHtml}
\title{Imports html files into a character vector.}
\usage{
ImportHtml(from = "", sample = NULL, nameOfProject = NULL,
  nameOfWebsite = NULL, pathToHtmlFolder = "", sort = TRUE,
  recursive = TRUE, includePath = FALSE)
}
\arguments{
\item{from}{Allows to choose which type of contents to import. Can be either "articles" or "index".}

\item{sample}{Defaults to NULL. If a numeric value n is provided, then instead of importing all html files it imports n random files.}

\item{nameOfProject}{Name of 'castarter' project. Must correspond to the name of a folder in the current working directory.}

\item{nameOfWebsite}{Name of a website included in a 'castarter' project. Must correspond to the name of a sub-folder of the project folder.}

\item{recursive}{Value is passed to list.files function when using the pathToHtmlFolder option.}
}
\value{
A character vector of html file, or a data frame if includePath is set to TRUE.
}
\description{
Imports html files already downloaded into a character vector.
}
\examples{
articlesHtml <- ImportHtml(from = articles, nameOfProject, nameOfWebsite)
}

