% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/prepareCorpus.R
\name{LoadAllDatasets}
\alias{LoadAllDatasets}
\title{Loads all datasets of a 'castarter' project.}
\usage{
LoadAllDatasets(nameOfProject)
}
\arguments{
\item{nameOfProject}{The name of the project whose datasets are to be imported.}
}
\value{
A data frame including all datasets of a project.
}
\description{
Takes all datasets from all websites in a project and outputs them in a data frame.
}
\examples{
allDatasets <- LoadAllDatasets(nameOfProject)
}

