% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareCorpus.R
\name{LoadDatasets}
\alias{LoadDatasets}
\title{Loads a specific set of datasets from existings 'castarter' projects.}
\usage{
LoadDatasets(projectsAndWebsites = NULL, type = "dataset",
  removeNAdates = TRUE, convertToUTF8 = FALSE,
  addProjectColumn = FALSE)
}
\arguments{
\item{projectsAndWebsites}{A character vector listing websites to be loaded in the format "project/website". If none is given, it will check if project and website have been set with `SetCastarter()`}

\item{type}{Defines the format in which the dataset will be loaded. Available options are:
\itemize{
 \item{"dataset"}{: Outputs a standard `castarter` dataset as data frame (a tibble). This is the default option.}
 \item{"datasetTidy"}{: Outputs a 'castarter' dataset, with one word per row, in line with tidy principles. See also the `tidytext` package.}
 \item{"datasetBySentence"}{: Outputs a 'castarter' dataset, with one sentence per row. This can be used to speed up the loading time of datasets analysed through the `AnalyseDataset()` function.}
}}

\item{removeNAdates}{Logical, defaults to TRUE. If TRUE, dataset items that do not have a date in the records are not imported.}

\item{convertToUTF8}{Logical, defaults to FALSE. If TRUE, converts the `title` and `text` column (or correspondent, in `datasetTidy` and `datasetBySentence`) to UTF8.}

\item{addProjectColumn}{Logical, defatults to FALSE. If TRUE, a column with the project name is appended.}
}
\value{
A data frame including all loaded datasets.
}
\description{
Takes a specific set of datasets from existings 'castarter' projects.
}
\examples{
\dontrun{
projectsAndWebsites <- c("ProjectX/Website1", "ProjectY/Website3", "ProjectZ/Website2")
allDatasets <- LoadDatasets(projectsAndWebsites)
}
}
