% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareCorpus.R
\name{LoadDatasets}
\alias{LoadDatasets}
\title{Loads a specific set of datasets from existings 'castarter' projects.}
\usage{
LoadDatasets(projectsAndWebsites, type = "dataset", removeNAdates = FALSE)
}
\arguments{
\item{projectsAndWebsites}{A character vector listing websites to be loaded in the format "nameOfProject/nameOfWebsite".}

\item{type}{Defines the format in which the dataset will be loaded. Available options are:
\itemize{
 \item{"dataset"}{: Outputs a 'castarter' dataset as data.frame. This is the default option.}
 \item{"corpus"}{: Outputs a corpus.}
}}
}
\value{
A data frame including all loaded datasets.
}
\description{
Takes a specific set of datasets from existings 'castarter' projects.
}
\examples{
projectsAndWebsites <- c("ProjectX/Website1", "ProjectY/Website3", "ProjectZ/Website2")
allDatasets <- LoadDatasets(projectsAndWebsites)
}

