% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preliminary.R
\name{LoadLatest}
\alias{LoadLatest}
\title{Finds the full path of the latest saved workspace for selected website.}
\usage{
LoadLatest(nameOfProject, nameOfWebsite)
}
\arguments{
\item{nameOfProject}{Name of 'castarter' project. Must correspond to the name of a folder in the current working directory.}

\item{nameOfWebsite}{Name of a website included in a 'castarter' project. Must correspond to the name of a sub-folder of the project folder.}
}
\description{
Used in combination with load, it loads the most recent workspace saved for a given website based on the date included in the filename of the .Rdata file.
}
\examples{
load(LoadLatest(nameOfProject, nameOfWebsite))
}

