% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{SetCastarter}
\alias{SetCastarter}
\title{Sets 'castarter' options}
\usage{
SetCastarter(nameOfProject = NULL, nameOfWebsite = NULL)
}
\arguments{
\item{nameOfProject}{Name of 'castarter' project. Must correspond to the name of a folder in the current working directory.}

\item{nameOfWebsite}{Name of a website included in a 'castarter' project. Must correspond to the name of a sub-folder of the project folder.}
}
\value{
Nothing. Used for its side effects (stores settings).
}
\description{
It allows to preliminary store options frequently used by 'castarter', thus removing the requirement to specify them each time a function is called.
}
\examples{
SetCastarter(nameOfProject, nameOfWebsite)
}

