% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/showMostFrequent.R
\name{ShowMostFrequent}
\alias{ShowMostFrequent}
\title{Shows most frequent terms in a corpus}
\usage{
ShowMostFrequent(corpusDtm, mode = "data.frame", number = 10,
  specificTerms = NULL, stemCompletion = FALSE, corpusOriginal = "",
  minFrequency = 0)
}
\arguments{
\item{corpusDtm}{A document term matrix.}

\item{mode}{Defines the type of output. Can be
\itemize{
 \item{"data.frame"}{: Outputs a data.frame This is the default option.}
 \item{"barchart"}{: Outputs a ggplot2 barchart.}
 \item{"wordcloud"}{: Outputs a wordcloud.}
}}

\item{specificTerms}{A character vector, defaults to NULL. If specificTerms is provided, only terms included in this vector will be included in the output.}
}
\value{
A data.frame, barchart, or wordcloud as defined with the 'mode' parameter.
}
\description{
Shows most frequent terms in a corpus starting from a document term matrix.
}
\examples{
corpusDtm <- DocumentTermMatrix(corpus).
mostFrequent <- ShowMostFrequent(corpusDtm)
}

