% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/showMostFrequent.R
\name{ShowMostFrequent}
\alias{ShowMostFrequent}
\title{Shows most frequent terms in a corpus}
\usage{
ShowMostFrequent(corpusDtm, mode = "vector", number = 10,
  stemCompletion = FALSE, corpusOriginal = "", minFrequency = 0)
}
\arguments{
\item{corpusDtm}{A document term matrix.}

\item{mode}{Defines the type of output. Can be
\itemize{
 \item{"vector"}{Outputs a named character vector.}
 \item{"barchart"}{Outputs a ggplot2 barchart. This is the default option.}
 \item{"wordcloud"}{Outputs a wordcloud.}
}}
}
\value{
A vector, barchart, or wordcloud as defined with the 'mode' parameter.
}
\description{
Shows most frequent terms in a corpus starting from a document term matrix.
}
\examples{
corpusDtm <- DocumentTermMatrix(corpus).
mostFrequent <- ShowMostFrequent(corpusDtm)
}

