% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.R
\name{SubsetDataset}
\alias{SubsetDataset}
\title{Subsets a 'castarter' dataset}
\usage{
SubsetDataset(
  dataset,
  terms = NULL,
  startDate = NULL,
  endDate = NULL,
  sample = NULL
)
}
\arguments{
\item{dataset}{A dataset created with 'castarter'.}

\item{terms}{A character vector. Only articles including one of these terms are included in the output.}

\item{startDate, }{endDate A date in the format YYYY-MM-DD.}
}
\value{
A dataset in the 'castarter' format.
}
\description{
Subsets a 'castarter' dataset
}
\examples{
\dontrun{
SubsetDataset(dataset, terms)
}
}
