% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateDataset.R
\name{UpdateDataset}
\alias{UpdateDataset}
\title{Updates a dataset}
\usage{
UpdateDataset(dataset, articlesLinks = NULL, numberOfIndexPages = 10,
  wget = FALSE, wait = 3, project = NULL, website = NULL)
}
\arguments{
\item{dataset}{A 'castarter' dataset.}

\item{project}{Name of 'castarter' project. Must correspond to the name of a folder in the current working directory.}

\item{website}{Name of a website included in a 'castarter' project. Must correspond to the name of a sub-folder of the project folder.}
}
\value{
A castarter dataset.
}
\description{
Updates a dataset given a previously created parameters export file.
}
\examples{
dataset <- UpdateDataset(dataset)
}

